feat: add no-plugin demo page and upgrade to v1.5.11
Split the example into with/without plugins multi-page demos, extract shared app bootstrapping, and refine Vite chunking so hosts can compare bare viewer vs plugin setups.
