+++ package.json
@@ -21,11 +21,11 @@
     "vite-plugin-static-copy": "^3.1.1"
   },
   "dependencies": {
-    "@mlightcad/cad-html-plugin": "1.5.10",
-    "@mlightcad/cad-pdf-plugin": "1.5.10",
-    "@mlightcad/cad-svg-plugin": "1.5.10",
-    "@mlightcad/cad-simple-ui-plugin": "1.5.10",
-    "@mlightcad/cad-simple-viewer": "1.5.10",
+    "@mlightcad/cad-html-plugin": "1.5.11",
+    "@mlightcad/cad-pdf-plugin": "1.5.11",
+    "@mlightcad/cad-svg-plugin": "1.5.11",
+    "@mlightcad/cad-simple-ui-plugin": "1.5.11",
+    "@mlightcad/cad-simple-viewer": "1.5.11",
     "@mlightcad/data-model": "1.12.5"
   }
 }
